requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
pandas==2.2.3
scikit-learn==1.5.2
numpy==1.26.4
django-filter==24.3
tqdm==2.2.3
ijson==3.3.0
